adding performance measurements to Cholesky
